backend/requirements.txt
fastapi>=0.104
uvicorn>=0.24
langchain>=0.3
langchain-openai>=0.3
python-dotenv>=1.0
langgraph>=0.2
pandas>=2.0
numpy>=1.24
scipy>=1.11
scikit-learn>=1.3
pysr>=0.19
python-multipart>=0.0.6
APScheduler>=3.10
